fix validation for country textbox as described in #130
Validation for country textbox of export tool is set valid if non-blank and matches value from config file.
Added warning if it is blank and error if it is non-blank but doesn't match config. This second condition can't currently be hit as the box is non-editable but added the logic in case that were to be changed.