Use current document, not MXD as saved on disk. Do not crash if single page export filename not created.
Changed logic of call to python code so that the current map document is used, so that the export works immediately after activating DDP and will reflect current map state, not last saved state.

Fixed check for exported filesize in python to not crash if filename doesn't match that expected (or doesn't exist at all due to some other failure!). This occurred when the mxd filename had a dot in it, arcpy only used part before first dot.